requirements.txt
Flask
tensorflow
numpy
Pillow
mysql-connector-python
